Create Project complete---Research Environment
Here is a collection of code written on the Quantopian research environment (Python 2.7 Jupyter notebook). eight of individual stock in a portfolio, SID of stock, portfolio past 5 months mean and standard deviation are returned as outputs into the Quantopian IDE. We manually tried through price data for each training-testing window and gathered the data for Jan to Oct in 2018.